Further restructure codebase
