+++ src/application/modules/composition/programs/showCompositionMap.frag
