+++ src/application/modules/composition/programs/showCompositionVectors.frag
