+++ src/application/modules/layers/programs/showLayerPoints.frag
