+++ src/application/modules/layers/programs/showLayerPoints.vert
