+++ src/application/modules/layers/programs/showTexture.ts
@@ -1,6 +1,9 @@
 import { type ScreenTranslation } from "@/dtos/application.dto";
+import {
+  createProgram,
+  type WebGLRenderer,
+} from "@/infrastructure/webgl/lib/webgl";
 
-import { createProgram, type WebGLRenderer } from "../lib/webgl";
 import textureFragmentShader from "./showTexture.frag";
 import textureVertexShader from "./showTexture.vert";
