+++ src/domain/animation/file1/convert.spec.ts
@@ -1,4 +1,5 @@
-import { v2Format } from "../file2/mockdata/file2.0";
+import { v2Format } from "@/domain/animation/file2/mockdata/file2.0";
+
 import { convertFromV2 } from "./convert";
 import { v1Format } from "./mockdata/file1.1";
