+++ src/domain/animation/file2/convert.spec.ts
@@ -1,6 +1,6 @@
+import { v1Format } from "@/domain/animation/file1/mockdata/file1.1";
 import { type ImageDefinition } from "@/dtos/animation-file1.dto";
 
-import { v1Format } from "../file1/mockdata/file1.1";
 import { convertFromV1 } from "./convert";
 import { v2Format } from "./mockdata/file2.0";
