+++ src/domain/animation/file2/verifyFile.spec.ts
@@ -1,4 +1,5 @@
-import { v1Format } from "../file1/mockdata/file1.1";
+import { v1Format } from "@/domain/animation/file1/mockdata/file1.1";
+
 import { v2Format } from "./mockdata/file2.0";
 import { verifyFile } from "./verifyFile";
