+++ src/ui/components/atoms/ToolSeparator/ToolSeparator.stories.tsx
@@ -1,9 +1,9 @@
 import preview from "@sb/preview";
 
+import { Icon } from "@/ui/components/atoms/Icon/Icon";
+import { ToolButton } from "@/ui/components/atoms/ToolButton/ToolButton";
 import { ToolBar } from "@/ui/components/molecules/ToolBar/ToolBar";
 
-import { Icon } from "../Icon/Icon";
-import { ToolButton } from "../ToolButton/ToolButton";
 import { ToolSeparator as ToolSeparatorComponent } from "./ToolSeparator";
 
 const meta = preview.meta({
