+++ src/ui/components/pages/Composition-v1.stories.tsx
@@ -1,6 +1,11 @@
 import preview from "@sb/preview";
 
 import {
+  storyTreeItems,
+  type ToolsProvider,
+} from "@/ui/components/organisms/Tree/storybookTreeDataProvider";
+
+import {
   Column,
   Control,
   ControlPanel,
@@ -29,10 +34,6 @@ import {
   Tree,
   TreeEnvironment,
 } from "..";
-import {
-  storyTreeItems,
-  type ToolsProvider,
-} from "../organisms/Tree/storybookTreeDataProvider";
 
 const meta = preview.meta({
   title: "Pages/Composition",
